* libmutil: Fixes to compile libmutil.dll under visual studio
git-svn-id: svn://svn.minisip.org/minisip/trunk@961 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmutil/include/libmutil/cert.h
@@ -148,7 +148,7 @@ class certificate_chain: public MObject{
 
 		void clear();
 
-		int length(){ return cert_list.size(); };
+		int length(){ return (int)cert_list.size(); };
 		void lock();
 		void unlock();
