+++ libmutil/source/cert.cxx
@@ -217,13 +217,13 @@ string certificate::get_cn(){
 	string name = get_name();
 	unsigned pos, pos2;
 
-	pos = name.find( "/CN=" );
+	pos = (unsigned int)name.find( "/CN=" );
 
 	if( pos == string::npos ){
 		return "No common name";
 	}
 
-	pos2 = name.find( "/", pos + 1 );
+	pos2 = (unsigned int)name.find( "/", pos + 1 );
 
 	return name.substr( pos + 4, pos2 - pos - 4 );
 }
@@ -239,13 +239,13 @@ string certificate::get_issuer_cn(){
 	string name = get_issuer();
 	unsigned pos, pos2;
 
-	pos = name.find( "/CN=" );
+	pos = (unsigned int)name.find( "/CN=" );
 
 	if( pos == string::npos ){
 		return "No common name";
 	}
 
-	pos2 = name.find( "/", pos + 1 );
+	pos2 = (unsigned int)name.find( "/", pos + 1 );
 
 	return name.substr( pos + 4, pos2 - pos - 4 );
 }
